# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification for RGI Strategic Intelligence System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunity-windows/config:
    get:
      operationId: getStoryOpportunityConfig
      tags: [opportunities]
      summary: Get internal Story Opportunity feature configuration
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      responses:
        "200":
          description: Feature configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryOpportunityConfigResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /opportunity-windows:
    get:
      operationId: listStoryOpportunityWindows
      tags: [opportunities]
      summary: List completed frozen Story Opportunity windows
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      responses:
        "200":
          description: Frozen window list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListStoryOpportunityWindowsResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /opportunity-windows/current:
    get:
      operationId: getCurrentStoryOpportunityWindow
      tags: [opportunities]
      summary: Get the latest completed frozen window and its shortlist
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      responses:
        "200":
          description: Current window or an explicit no-window response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryOpportunityListResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /opportunity-windows/calculate:
    post:
      operationId: calculateStoryOpportunityWindow
      tags: [opportunities]
      summary: Explicitly and idempotently calculate one frozen daily window
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateStoryOpportunityWindowBody"
      responses:
        "200":
          description: Existing frozen window returned unchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculateStoryOpportunityWindowResponse"
        "201":
          description: Frozen window created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculateStoryOpportunityWindowResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /opportunity-windows/{windowId}/opportunities:
    get:
      operationId: listStoryOpportunitiesForWindow
      tags: [opportunities]
      summary: List the frozen shortlist for one window
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - $ref: "#/components/parameters/StoryOpportunityWindowId"
      responses:
        "200":
          description: Frozen opportunity shortlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryOpportunityListResponse"
        "404":
          description: Window not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /story-opportunities/{id}:
    get:
      operationId: getStoryOpportunity
      tags: [opportunities]
      summary: Get one Story Opportunity workbench record
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - $ref: "#/components/parameters/StoryOpportunityId"
      responses:
        "200":
          description: Story Opportunity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryOpportunityDetailResponse"
        "404":
          description: Opportunity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /story-opportunities/{id}/matches:
    get:
      operationId: listStoryOpportunityProfessorMatches
      tags: [opportunities]
      summary: List persisted deterministic Professor Matches
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - $ref: "#/components/parameters/StoryOpportunityId"
      responses:
        "200":
          description: Professor Match list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorMatchListResponse"
        "404":
          description: Opportunity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /story-opportunities/{id}/select-professor:
    post:
      operationId: selectStoryOpportunityProfessor
      tags: [opportunities]
      summary: Manually select or change the professor for an opportunity
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - $ref: "#/components/parameters/StoryOpportunityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectStoryOpportunityProfessorBody"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryOpportunityDetailResponse"
        "400":
          description: Override reason required or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Opportunity revision conflict, frozen match exclusion, or PROFESSOR_PROFILE_REVISION_CONFLICT when the current Professor Profile no longer matches the frozen revision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /story-opportunities/{id}/clear-professor:
    post:
      operationId: clearStoryOpportunityProfessor
      tags: [opportunities]
      summary: Clear the current manual professor selection
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - $ref: "#/components/parameters/StoryOpportunityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClearStoryOpportunityProfessorBody"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryOpportunityDetailResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "409":
          description: Opportunity revision conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /story-opportunities/{id}/update-angle:
    post:
      operationId: updateStoryOpportunityAngle
      tags: [opportunities]
      summary: Update the editor-controlled recommended angle
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - $ref: "#/components/parameters/StoryOpportunityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoryOpportunityAngleBody"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryOpportunityDetailResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "409":
          description: Opportunity revision conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
...
components:
  securitySchemes:
    FirebaseEditorAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Firebase ID token for a user whose UID is present in the server-only approved RGI editor allowlist. Missing verification configuration fails closed.
    InternalServiceKey:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Trusted server-to-server operational credential. This key must never be embedded in or sent by the browser application.
  responses:
    InternalEditorUnauthorized:
      description: Missing, malformed, invalid, expired, or revoked Firebase ID token, or invalid internal service credential
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalEditorForbidden:
      description: Authenticated user is not an approved RGI editor, or the requested feature is disabled
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalEditorAuthUnavailable:
      description: Server-side Firebase token verification or approved-editor configuration is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    StoryOpportunityId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{8,160}$'
    StoryOpportunityWindowId:
      name: windowId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{8,160}$'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        retryable:
          type: boolean
        userMessage:
          type: string
      required:
        - error
...
    StoryOpportunityConfigResponse:
      type: object
      properties:
        readsEnabled:
          type: boolean
        writesEnabled:
          type: boolean
      required: [readsEnabled, writesEnabled]
    StoryOpportunityWindow:
      type: object
      properties:
        id: { type: string }
        snapshotRevision: { type: integer, minimum: 1 }
        asOf: { type: string, format: date-time }
        windowStart: { type: string, format: date-time }
        windowEnd: { type: string, format: date-time }
        operationalTimezone: { type: string }
        localCutoff: { type: string }
        calculatedAt: { type: string, format: date-time }
        status: { type: string, enum: [completed] }
        configurationVersion: { type: string }
        windowAlgorithmVersion: { type: string }
        selectionAlgorithmVersion: { type: string }
        scoringVersion: { type: string }
        normalizationVersion: { type: string }
        taxonomyVersion: { type: string }
        matchingAlgorithmVersion: { type: string }
        coverageCalculationVersion: { type: string }
        minimumNormalizedRelevance: { type: number }
        maximumOpportunities: { type: integer }
        maximumPerSource: { type: integer }
        maximumPerPrimaryTopic: { type: integer }
        lowCoverageWarningThreshold: { type: number }
        totalArticlesConsidered: { type: integer }
        eligibleArticleCount: { type: integer }
        qualifyingArticleCount: { type: integer }
        opportunityCount: { type: integer }
        fallbackTimestampCount: { type: integer }
      required:
        - id
        - snapshotRevision
        - asOf
        - windowStart
        - windowEnd
        - operationalTimezone
        - localCutoff
        - calculatedAt
        - status
        - configurationVersion
        - windowAlgorithmVersion
        - selectionAlgorithmVersion
        - scoringVersion
        - normalizationVersion
        - taxonomyVersion
        - matchingAlgorithmVersion
        - coverageCalculationVersion
        - minimumNormalizedRelevance
        - maximumOpportunities
        - maximumPerSource
        - maximumPerPrimaryTopic
        - lowCoverageWarningThreshold
        - totalArticlesConsidered
        - eligibleArticleCount
        - qualifyingArticleCount
        - opportunityCount
        - fallbackTimestampCount
...
    StoryOpportunity:
      type: object
      properties:
        id: { type: string }
        revision: { type: integer }
        windowId: { type: string }
        windowStart: { type: string, format: date-time }
        windowEnd: { type: string, format: date-time }
        operationalTimezone: { type: string }
        primaryArticleId: { type: integer }
        primaryEvidence: { $ref: "#/components/schemas/StoryOpportunityEvidence" }
        supportingEvidence:
          type: array
          items: { $ref: "#/components/schemas/StoryOpportunityEvidence" }
        sourceName: { type: string }
        canonicalUrl: { type: string }
        effectivePublishedAt: { type: string, format: date-time }
        timestampSource: { type: string, enum: [publishedAt, scrapedAt] }
        timestampFallback: { type: boolean }
        originalRgiRelevanceScore: { type: number }
        originalRgiRelevanceScale: { type: string, enum: [1-10] }
        originalRgiRelevanceField: { type: string, enum: [relevancyScore] }
        normalizedRgiRelevanceScore: { type: number, minimum: 0, maximum: 100 }
        relevanceExplanation: { type: [string, "null"] }
        relevanceComponents:
          type: [object, "null"]
          additionalProperties: true
        relevanceScoringVersion: { type: string }
        relevanceNormalizationVersion: { type: string }
        sourceAuthorityScore: { type: number }
        primaryTopic: { type: string }
        primaryTopicLabel: { type: string }
        normalizedTopics:
          type: array
          items: { type: string }
        unknownTaxonomyTerms:
          type: array
          items: { type: string }
        discipline: { type: [string, "null"] }
        industries:
          type: array
          items: { type: string }
        regions:
          type: array
          items: { type: string }
        entities:
          type: array
          items: { type: string }
        recommendedAngle: { type: string, maxLength: 500 }
        shortlistPosition: { type: integer, minimum: 1 }
        selectionConfiguration: { $ref: "#/components/schemas/StoryOpportunitySelectionConfiguration" }
        professorMatches:
          type: array
          items: { $ref: "#/components/schemas/ProfessorMatch" }
        selectedProfessor:
          anyOf:
            - $ref: "#/components/schemas/ProfessorSelection"
            - type: "null"
        selectionHistory:
          type: array
          items: { $ref: "#/components/schemas/ProfessorSelectionHistoryEntry" }
        workflowState: { type: string, enum: [shortlisted, professor_selected, closed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        configurationVersion: { type: string }
        taxonomyVersion: { type: string }
        matchingAlgorithmVersion: { type: string }
      required:
        - id
        - revision
        - windowId
        - windowStart
        - windowEnd
        - operationalTimezone
        - primaryArticleId
        - primaryEvidence
        - supportingEvidence
        - sourceName
        - canonicalUrl
        - effectivePublishedAt
        - timestampSource
        - timestampFallback
        - originalRgiRelevanceScore
        - originalRgiRelevanceScale
        - originalRgiRelevanceField
        - normalizedRgiRelevanceScore
        - relevanceExplanation
        - relevanceComponents
        - relevanceScoringVersion
        - relevanceNormalizationVersion
        - sourceAuthorityScore
        - primaryTopic
        - primaryTopicLabel
        - normalizedTopics
        - unknownTaxonomyTerms
        - discipline
        - industries
        - regions
        - entities
        - recommendedAngle
        - shortlistPosition
        - selectionConfiguration
        - professorMatches
        - selectedProfessor
        - selectionHistory
        - workflowState
        - createdAt
        - updatedAt
        - configurationVersion
        - taxonomyVersion
        - matchingAlgorithmVersion
    ListStoryOpportunityWindowsResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/StoryOpportunityWindow" }
        total: { type: integer }
        readsEnabled: { type: boolean }
        writesEnabled: { type: boolean }
      required: [items, total, readsEnabled, writesEnabled]
    StoryOpportunityListResponse:
      type: object
      properties:
        window:
          anyOf:
            - $ref: "#/components/schemas/StoryOpportunityWindow"
            - type: "null"
        items:
          type: array
          items: { $ref: "#/components/schemas/StoryOpportunity" }
        total: { type: integer }
        readsEnabled: { type: boolean }
        writesEnabled: { type: boolean }
      required: [window, items, total, readsEnabled, writesEnabled]
    StoryOpportunityDetailResponse:
      type: object
      properties:
        opportunity: { $ref: "#/components/schemas/StoryOpportunity" }
        readsEnabled: { type: boolean }
        writesEnabled: { type: boolean }
      required: [opportunity, readsEnabled, writesEnabled]
    ProfessorMatchListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ProfessorMatch" }
        total: { type: integer }
      required: [items, total]
    CalculateStoryOpportunityWindowBody:
      type: object
      additionalProperties: false
      properties:
        asOf: { type: string, format: date-time }
        snapshotRevision: { type: integer, minimum: 1, default: 1 }
      required: [asOf]
    CalculateStoryOpportunityWindowResponse:
      type: object
      properties:
        created: { type: boolean }
        window: { $ref: "#/components/schemas/StoryOpportunityWindow" }
        opportunities:
          type: array
          items: { $ref: "#/components/schemas/StoryOpportunity" }
        readsEnabled: { type: boolean }
        writesEnabled: { type: boolean }
      required: [created, window, opportunities, readsEnabled, writesEnabled]
    SelectStoryOpportunityProfessorBody:
      type: object
      additionalProperties: false
      properties:
        professorId:
          type: string
          pattern: '^[A-Za-z0-9_-]{8,128}$'
        reason:
          type: string
          maxLength: 1000
        expectedRevision:
          type: integer
          minimum: 1
      required: [professorId, expectedRevision]
...
    ClearStoryOpportunityProfessorBody:
      type: object
      additionalProperties: false
      properties:
        expectedRevision:
          type: integer
          minimum: 1
        reason:
          type: string
          maxLength: 1000
      required: [expectedRevision]
    UpdateStoryOpportunityAngleBody:
      type: object
      additionalProperties: false
      properties:
        expectedRevision:
          type: integer
          minimum: 1
        angle:
          type: string
          minLength: 1
          maxLength: 500
      required: [expectedRevision, angle]
...